pyproject.toml
[project]
name = "qr-code-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "qrcode-artistic>=3.0.2",
    "sengo>=0.0.2",
]
